Fixed bad message construction and merged the createpage-login-href and createpage-login-required2 i18n messages into the main createpage-login-required message
Tested and bugfixed the whole "anonymous user tries to upload an image via a createplate that has an infobox" scenario.
Now when an anon attempts to use an "Insert image" button, the error message is correctly shown and clicking on the "log in" link brings up the jQuery UI dialog instead of taking the user directly to Special:UserLogin.

Change-Id: I930ba7b73a2022d2a42ca57bbbddb825abd24c0d