"Fix" the apihelp i18n messages
Per post-merge review at Change-Id: I8c1bce43bbb11c84c047df6c02af305e79d469e6.

This is more of a description of the issue at hand than anything useful for an end-user, but oh well.
The whole createapage-upload API module is literally not anything even remotely reusable and it's not really meant to be.
Rather, its whole existence is an unfortunate thing and at some point, when someone has enough time on a rainy day, the whole upload code should be rewritten to use core classes and code in a more reasonable manner.
Until that happens, these hacky API modules introduced in the aforementioned changeset as a replacement for the old $wgAjaxExportList code will have to stay, alas.

Either way, "qqq as English" is not useful so perhaps this little CTA will encourage someone to take action and look into reworking the upload code in this extension...

Change-Id: Ic843eec91583c53f13ecd687640dd0abc423b548